Beginnings of API v1
Move shared code into top level application controller, add Devices#show
and Devices#index using new presentations
