+++ app/controllers/v1/application_controller.rb
@@ -1,0 +1,4 @@
+module V1
+  class ApplicationController < ::ApplicationAPIController
+  end
+end
