--- app/views/v001/devices/current_user_index.jbuilder
@@ -1,1 +1,0 @@
-json.array! @devices, partial: 'device', as: :device
