--- app/views/v001/devices/index.jbuilder
@@ -1,3 +1,0 @@
-json.devices do
-  json.array! @devices, partial: 'device', as: :device
-end
